requirements.txt
transformers
langchain
langchain-huggingface
torch 
torchvision 
torchaudio
streamlit
python-dotenv
accelerate
sentencepiece
